Implement sanitization validation for POST, PUT, and DELETE requests

+++ backend/controllers/postsController.js
@@ -1,6 +1,5 @@
 const asyncHandler = require('express-async-handler');
 const { body, validationResult } = require('express-validator');
-const bcrypt = require('bcryptjs');
 
 const { APIError } = require('../utils/helpers');
 const Post = require('../models/post');
@@ -15,23 +14,40 @@ exports.postsGet = asyncHandler(async (req, res) => {
   }
 });
 
-exports.postPost = asyncHandler(async (req, res) => {
-  const postInfo = {
-    user: req.user,
-    title: req.body.title,
-    content: req.body.content,
-    isPublished: req.body.isPublished,
-  };
-  try {
-    const post = await Post.create(postInfo);
-    await User.findByIdAndUpdate(req.user._id, {
-      posts: [...req.user.posts, post],
-    });
-    return res.json({ status: 'success', data: post });
-  } catch (err) {
-    throw APIError(err.status, err.message, 'database_error');
-  }
-});
+exports.postPost = [
+  body('title')
+    .trim()
+    .isLength({ max: 500 })
+    .withMessage('Title must be within 500 characters')
+    .escape(),
+  body('content')
+    .trim()
+    .isLength({ min: 1, max: 10000 })
+    .withMessage('Content must be within 1—10000 characters')
+    .escape(),
+
+  asyncHandler(async (req, res) => {
+    const errors = validationResult(req).array();
+    if (errors.length) throw APIError(400, errors[0].msg, 'invalid_input');
+
+    const postInfo = {
+      user: req.user,
+      title: req.body.title || undefined,
+      content: req.body.content,
+      isPublished: req.body.isPublished,
+    };
+
+    try {
+      const post = await Post.create(postInfo);
+      await User.findByIdAndUpdate(req.user._id, {
+        posts: [...req.user.posts, post],
+      });
+      return res.json({ status: 'success', data: post });
+    } catch (err) {
+      throw APIError(err.status, err.message, 'database_error');
+    }
+  }),
+];
 
 exports.postGet = asyncHandler(async (req, res) => {
   try {
@@ -43,27 +59,59 @@ exports.postGet = asyncHandler(async (req, res) => {
   }
 });
 
-exports.postPut = asyncHandler(async (req, res) => {
-  const post = await Post.findById(req.params.postID).exec();
-  if (!post) throw APIError(404, 'Post not found', 'resource_not_found');
-  const update = {
-    user: req.user,
-    title: req.body.title,
-    content: req.body.content,
-    isPublished: req.body.isPublished,
-    _id: req.params.postID,
-  };
-  try {
-    const data = await Post.findByIdAndUpdate(req.params.postID, update);
-    return res.json({ status: 'success', data });
-  } catch (err) {
-    throw APIError(err.status, err.message, 'database_error');
-  }
-});
+exports.postPut = [
+  body('title')
+    .trim()
+    .isLength({ max: 500 })
+    .withMessage('Title must be within 500 characters')
+    .escape(),
+  body('content')
+    .trim()
+    .isLength({ min: 1, max: 10000 })
+    .withMessage('Content must be within 1—10000 characters')
+    .escape(),
+
+  asyncHandler(async (req, res) => {
+    const errors = validationResult(req).array();
+    if (errors.length) throw APIError(400, errors[0].msg, 'invalid_input');
+
+    const post = await Post.findById(req.params.postID).exec();
+    if (!post) throw APIError(404, 'Post not found', 'resource_not_found');
+    if (post.user._id.toString() !== req.user._id.toString())
+      throw APIError(
+        401,
+        'User is unauthorized to update this post',
+        'unauthorized',
+      );
+
+    const update = {
+      user: req.user,
+      title: req.body.title,
+      content: req.body.content,
+      isPublished: req.body.isPublished,
+      _id: req.params.postID,
+    };
+
+    try {
+      const data = await Post.findByIdAndUpdate(req.params.postID, update);
+      return res.json({ status: 'success', data });
+    } catch (err) {
+      throw APIError(err.status, err.message, 'database_error');
+    }
+  }),
+];
 
 exports.postDelete = asyncHandler(async (req, res) => {
   const post = await Post.findById(req.params.postID).exec();
   if (!post) throw APIError(404, 'Post not found', 'resource_not_found');
+
+  if (post.user._id.toString() !== req.user._id.toString()) {
+    throw APIError(
+      401,
+      'User is unauthorized to delete this post',
+      'unauthorized',
+    );
+  }
 
   try {
     const data = await Post.findByIdAndDelete(req.params.postID);
